Update SARS-CoV-2 genomes and add a new SARS-CoV-2_loose-surrounding FASTA
SARS-CoV-2_loose-surrounding.fasta.gz contains everything in the
SARS-related species, except for SARS-CoV-2 and two sequences (RaTG13
and a pangolin sequence) that are very closely related to SARS-CoV-2.
This lets us design for SARS-CoV-2 in a way where we can be specific but
without constraining the design too much (because those two sequences
can really constrain the design).