Re-run SARS-CoV-2-related and SARS species designs with subsampled SARS-CoV-2 genomes
These two taxa combine SARS-CoV-2 with other taxa. When there are ~10k
SARS-CoV-2 genomes, this overwhelms the calculation of expected activity
and the other taxa can easily be ignored (e.g., guides may not detect
SARS-CoV-1 in the SARS species, and only detect SARS-CoV-2). To avoid
this issue we could weight the calculation of expected activity
differently depending on subspecies (to balance SARS-CoV-1 and -2). This
simply subsamples SARS-CoV-2 genomes to get around this issue.

It uses a number of SARS-CoV-2 genomes on a similar order as the number
of SARS-CoV-1 for the SARS species design. For the SARS-CoV-2-related
design, it uses a much smaller number: similar to the number of
non-SARS-CoV-2 genomes within the taxon.