Add Enterovirus B designs for experimental testing
This commit adds a script to design only for Enterovirus B. It adjusts
some parameters compared to the default. It also adds a script to design
test targets, against with adjusted parameters owing to the extremely
high diversity of the species.

The commit also adds the output of these scripts.